+++ src/apis/StorageController.php
@@ -18,10 +18,11 @@ use luya\helpers\FileHelper;
 use yii\web\BadRequestHttpException;
 use yii\web\NotFoundHttpException;
 use luya\admin\models\StorageImage;
-use luya\admin\file\Item;
-use luya\helpers\ArrayHelper;
 use yii\data\ActiveDataProvider;
 use luya\admin\models\TagRelation;
+use luya\admin\traits\TaggableTrait;
+use luya\admin\storage\BaseFileSystemStorage;
+use luya\admin\events\FileEvent;
 
 /**
  * Filemanager and Storage API.
@@ -120,7 +121,7 @@ class StorageController extends RestController
             throw new NotFoundHttpException("Unable to find the given file to toggle the tag.");
         }
     
-        $relation = TagRelation::find()->where(['table_name' => StorageFile::tableName(), 'pk_id' => $fileId, 'tag_id' => $tagId])->one();
+        $relation = TagRelation::find()->where(['table_name' => TaggableTrait::cleanBaseTableName(StorageFile::tableName()), 'pk_id' => $fileId, 'tag_id' => $tagId])->one();
 
         if ($relation) {
             $relation->delete();
@@ -129,7 +130,7 @@ class StorageController extends RestController
         }
 
         $model = new TagRelation();
-        $model->table_name = StorageFile::tableName();
+        $model->table_name = TaggableTrait::cleanBaseTableName(StorageFile::tableName());
         $model->pk_id = $fileId;
         $model->tag_id = $tagId;
 
@@ -257,6 +258,9 @@ class StorageController extends RestController
         $model->attributes = $post;
         
         if ($model->update(true, ['name_original', 'inline_disposition']) !== false) {
+
+            Yii::$app->storage->trigger(BaseFileSystemStorage::FILE_UPDATE_EVENT, new FileEvent(['file' => $model]));
+            
             $this->flushApiCache($model->folder_id, $pageId);
             return $model;
         }
@@ -356,7 +360,11 @@ class StorageController extends RestController
                 
                 if (Storage::replaceFile($file->systemFileName, $newFileSource, $raw['name'])) {
                     foreach (StorageImage::find()->where(['file_id' => $file->id])->all() as $img) {
-                        $removal = Storage::removeImage($img->id, false);
+                        // remove the source
+                        if ($img->deleteSource()) {
+                            // recreate image filters
+                            $img->imageFilter($img->filter_id, false);
+                        }
                     }
                     
                     // calculate new file files based on new file