+++ src/ngrest/plugins/SelectModel.php
@@ -156,7 +156,7 @@ class SelectModel extends Select
         
         $values = [];
         foreach ($defintion as $field) {
-            $data = $model->$field;
+            $data = $model->i18nAttributeValue($field);
             
             if (is_array($data)) {
                 $data = reset($data);
@@ -238,7 +238,7 @@ class SelectModel extends Select
     {
         return [
             $this->createCrudLoaderTag($this->modelClass, $ngModel),
-            $this->createFormTag('zaa-select', $id, $ngModel, ['initvalue' => $this->initValue, 'options' => $this->getServiceName('selectdata')]),
+            $this->createFormTag('zaa-select', $id, $ngModel, array_filter(['initvalue' => $this->initValue, 'options' => $this->getServiceName('selectdata')])),
         ];
     }