+++ src/views/aws/groupauth/index.php
@@ -13,8 +13,8 @@ use luya\admin\Module;
                 <span class="material-icons card-toggle-indicator">keyboard_arrow_down</span>
                 {{ name }}
             </div>
-            <div class="table-responsive-wrapper" ng-show="groupVisibility">
-                <table class="table m-0 p-0">
+            <div class="table-responsive" ng-show="groupVisibility">
+                <table class="table p-0">
                     <thead>
                         <tr>
                             <th class="w-25">